ailog-bot/requirements.txt
Flask==1.1.2
line-bot-sdk
werkzeug
boto3
Werkzeug
pytz
validators
urlextract
